requirements.txt
flask>=3.0
flask-login>=0.6
python-dotenv>=1.0
waitress>=3.0
